most needed REST API for UI and #27 done

+++ scripts/server.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 usage() {
-  echo "Usage qsm server [term, launch, test, stop, stoptest]"
+  echo "Usage qsm server [term, launch, stop]"
   exit 1
 }
 
@@ -20,46 +20,45 @@ if [[ $? -ne 0 ]]; then
   exit 2
 fi
 
-port="8063"
-isTest=false
+backendDir="${rootDir}/backend"
 
-analyzeParameters() {
-  local portParam
-  port="8063"
-  for arg in "$@"; do
-    if [[ "$arg" == "-test" ]]; then
-      isTest=true
-      port="8877"
-      break
+getServerPort() {
+  local envFile
+  local port
+  envFile="${backendDir}/.env"
+  if [[ -e "$envFile" ]]; then
+    port=$(grep "SERVER_PORT" "$envFile" | awk -F '=' '{print $2}')
+    if [[ $? -ne 0 ]]; then
+      echo "ERROR: no SERVER_PORT found in $envFile"
+      exit 11
     fi
-  done
-  portParam=false
-  for arg in "$@"; do
-    if ${portParam}; then
-      port=arg
-      break
-    fi
-    if [[ "$arg" == "-port" ]]; then
-      portParam=true
-    fi
-  done
+    echo "$port"
+  else
+    echo "ERROR: The file $envFile does not exists"
+    exit 22
+  fi
 }
 
 getPidFile() {
-  if ${isTest}; then
-    echo "${logDir}/backend-test-${port}.pid"
-  else
-    echo "${logDir}/backend-${port}.pid"
+  local port
+  port="$(getServerPort)"
+  if [[ $? -ne 0 ]]; then
+    echo "$port"
+    exit 14
   fi
+  echo "${logDir}/backend-${port}.pid"
 }
 
 getLogFile() {
-  local outLog
-  if ${isTest}; then
-    outLog="${logDir}/backend-test-${port}.log"
-  else
-    outLog="${logDir}/backend-${port}.log"
+  local port
+  port="$(getServerPort)"
+  if [[ $? -ne 0 ]]; then
+    echo "$port"
+    exit 13
   fi
+
+  local outLog
+  outLog="${logDir}/backend-${port}.log"
   if [[ -e "$outLog" ]]; then
     mv "$outLog" "${outLog:0:-4}-$(date "+%Y-%m-%d_%H-%M-%S").log"
   fi
@@ -67,32 +66,56 @@ getLogFile() {
 }
 
 stopServer() {
+  local port
+  port="$(getServerPort)"
+  if [[ $? -ne 0 ]]; then
+    echo "$port"
+    exit 13
+  fi
+
   local pidFile
   local pidValue
-  analyzeParameters "$@"
   pidFile="$(getPidFile)"
+  if [[ $? -ne 0 ]]; then
+    echo "$pidFile"
+    exit 15
+  fi
+
   if [[ -e "$pidFile" ]]; then
     pidValue=$(cat "${pidFile}")
-    echo "Stopping server at pid $pidValue for port $port"
+    echo "INFO: Stopping server at pid $pidValue for port $port"
     # shellcheck disable=SC2086
     kill ${pidValue}
     sleep 1
     rm "${pidFile}"
+  else
+    echo "DEBUG: Server for port $port already stopped"
   fi
 }
 
 runServer() {
+  local pidFile
   local logFile
   local backendPid
-  analyzeParameters "$@"
+
   pidFile="$(getPidFile)"
+  if [[ $? -ne 0 ]]; then
+    echo "$pidFile"
+    exit 16
+  fi
   if [[ -e "$pidFile" ]]; then
     echo "Server on  port $port already running"
     exit 2
   fi
+
   logFile="$(getLogFile)"
+  if [[ $? -ne 0 ]]; then
+    echo "$pidFile"
+    exit 16
+  fi
+
   echo "INFO: Launching backend QSM log file at=$logFile"
-  cd "${rootDir}/backend" && ${go_exe} build || exit $?
+  cd "${backendDir}" && ${go_exe} build || exit $?
   nohup ./backend server "$@" 1>"${logFile}" 2>&1 &
   backendPid="$!"
   echo "INFO: Backend QSM launched PID=$backendPid"
@@ -101,7 +124,7 @@ runServer() {
 }
 
 runBackend() {
-  cd "${rootDir}/backend" && ${go_exe} build && ./backend server "$@"
+  cd "${backendDir}" && ${go_exe} build && ./backend server "$@"
 }
 
 commandName=$1
@@ -114,14 +137,8 @@ term)
 launch)
   runServer "$@"
   ;;
-test)
-  runServer -test "$@"
-  ;;
 stop)
   stopServer "$@"
-  ;;
-stoptest)
-  stopServer -test "$@"
   ;;
 *)
   usage
